# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/movies": {
            "get": {
                "description": "Get all movies from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get a list of movies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Movie"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new movie",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Create a new movie",
                "parameters": [
                    {
                        "description": "Title of the movie",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Watched status of the movie",
                        "name": "watched",
                        "in": "body",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all movies",
                "tags": [
                    "movies"
                ],
                "summary": "Delete all movies",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "post": {
                "description": "Set a movie as watched",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Set movie as watched",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "description": "Delete a movie by ID",
                "tags": [
                    "movies"
                ],
                "summary": "Delete a movie by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Movie": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "ID will be automatically generated when we insert a new Movie\nbson can also be represented as json so can define rules for it too",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "watched": {
                    "type": "boolean"
                }
            }
        }
    }
}